# chore: tweak deletion batch size & log messages